♻️ starting move from util to metrics folder
