+++ src/clients/RenewableEnergy.tsx
@@ -8,7 +8,7 @@ import config, { RenewableResults } from "../_config";
 import { Collapse } from "../components/Collapse";
 import { ClassTable } from "../components/ClassTable";
 import SketchClassTable from "../components/SketchClassTable";
-import { flattenSketchAllClass } from "../util/clientMetrics";
+import { flattenSketchAllClass } from "../metrics/clientMetrics";
 
 const CONFIG = config.renewable;
